mv check contract name to common package
