+++ contract/kernel/deploy.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/xuperchain/xuperunion/common"
 	"github.com/xuperchain/xuperunion/contract/wasm"
-	"github.com/xuperchain/xuperunion/permission/acl"
 	"github.com/xuperchain/xuperunion/permission/acl/utils"
 )
 
@@ -23,7 +23,7 @@ func (dm *DeployMethod) Invoke(ctx *KContext, args map[string][]byte) ([]byte, e
 		return nil, errors.New("invoke DeployMethod error, account name or contract name is nil")
 	}
 	// check if contractName is ok
-	if contractErr := acl.ValidContractName(string(contractName)); contractErr != nil {
+	if contractErr := common.ValidContractName(string(contractName)); contractErr != nil {
 		return nil, fmt.Errorf("deploy failed, contract `%s` contains illegal character, error: %s", contractName, contractErr)
 	}
 	_, err := ctx.ModelCache.Get(utils.GetAccountBucket(), accountName)
